removed obsolete code and renamed for clarity
